+++ go/cmd/vtgr/main.go
@@ -15,15 +15,21 @@ package main
 
 import (
 	"context"
+	"fmt"
 
 	"github.com/spf13/pflag"
 
 	"vitess.io/vitess/go/acl"
+	"vitess.io/vitess/go/vt/log"
 	"vitess.io/vitess/go/vt/servenv"
 	"vitess.io/vitess/go/vt/vtgr"
 )
 
+const deprecationMsg = "vtgr is deprecated and will be removed in Vitess 18. We recommend using VTOrc with semi-sync replication instead."
+
 func main() {
+	fmt.Println(deprecationMsg)
+
 	var clustersToWatch []string
 	servenv.OnParseFor("vtgr", func(fs *pflag.FlagSet) {
 		fs.StringSliceVar(&clustersToWatch, "clusters_to_watch", nil, `Comma-separated list of keyspaces or keyspace/shards that this instance will monitor and repair. Defaults to all clusters in the topology. Example: "ks1,ks2/-80"`)
@@ -31,6 +37,8 @@ func main() {
 		acl.RegisterFlags(fs)
 	})
 	servenv.ParseFlags("vtgr")
+
+	log.Warning(deprecationMsg)
 
 	// openTabletDiscovery will open up a connection to topo server
 	// and populate the tablets in memory